Add optional variable `AVRDUDE_MCU` to avrdude.mk
For some controllers, avrdude needs another spelling than the compiler
needs. For example the ATmega48PA. The compiler is fine with
`atmega48pa` but avrdude needs `m48p`.
This change lets the avrdude target use the new variable `AVRDUDE_MCU`
which is set to `MCU` by default. But the user is able to override it.